+++ changelog.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [2.6.5-dev]
+
 ## [2.6.4]
 
 ### enhancements